% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq_first_names.R
\docType{data}
\name{freq_first_names}
\alias{freq_first_names}
\title{Frequent Russian First Names}
\format{A data table with \Sexpr{nrow(rulexicon::freq_first_names)} rows
and \Sexpr{ncol(rulexicon::freq_first_names)} variables:
\describe{
    \item{name}{a first name}
    \item{freq}{the proportion relative to the most frequent name within the sex}
    \item{sex}{the sex corresponding to the name}
}}
\source{
\url{https://data.mos.ru/opendata/7704111479-svedeniya-o-naibolee-populyarnyh-mujskih-imenah-sredi-novorojdennyh}
\url{https://data.mos.ru/opendata/7704111479-svedeniya-o-naibolee-populyarnyh-jenskih-imenah-sredi-novorojdennyh}
}
\usage{
data(freq_first_names)
}
\description{
A dataset containing the most frequent Russian first names in Moscow region from 01.2015 to 03.2020.
}
\details{
TODO
}
\section{License}{
 See \url{https://apidata.mos.ru/Home/Terms}
}

\keyword{datasets}
