% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq_last_names.R
\docType{data}
\name{freq_last_names}
\alias{freq_last_names}
\title{Frequent Russian Last Names}
\format{A data table with \Sexpr{nrow(rulexicon::freq_last_names)} rows
and \Sexpr{ncol(rulexicon::freq_last_names)} variables:
\describe{
    \item{surname}{a last name}
    \item{freq}{the proportion relative to the most frequent surname}
}}
\source{
\url{http://imja.name/familii/pyatsot-chastykh-familij.shtml}
}
\usage{
data(freq_last_names)
}
\description{
A dataset containing the \Sexpr{nrow(rulexicon::freq_last_names)} most frequent Russian last names of Zhuravlev (2005).
}
\section{License}{

The dictionary is published under Public domain license.
}

\references{
Zhuravlev, A.F. (2005). On the Russian Family Names Statistics. Problems of Onomastics 2, pp. 126-146. URL: \url{http://hdl.handle.net/10995/1929}.
}
\keyword{datasets}
