% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hash_rusentilex_2016.R
\docType{data}
\name{hash_rusentilex_2016}
\alias{hash_rusentilex_2016}
\title{General Russian Sentiment Lexicon RuSentiLex 2016}
\format{A data table with \Sexpr{nrow(rulexicon::hash_rusentilex_2016)} rows and
\Sexpr{ncol(rulexicon::hash_rusentilex_2016)} variables:
\describe{
    \item{token}{the textual token (word or phrase)}
    \item{speech.part}{part of speech or type of syntactic group}
    \item{lemma}{initial word (phrase) in a lemmatized form}
    \item{sentiment}{one of positive, negative, neutral or positive/negative (indefinite, depends on the context)}
    \item{source}{opinion, feeling (private state), or fact (sentiment connotation)}
    \item{ambiguity}{if sentiment is different for senses of an ambiguous word, then sentiment orientations for all senses
    are described, the senses are labeled with the RuThes concept names}
}}
\source{
\url{https://www.labinform.ru/pub/rusentilex/rusentilex_2016.txt}
}
\usage{
data(hash_rusentilex_2016)
}
\description{
The original table of general Russian sentiment lexicon - RuSentiLex, version 2016.
 Table contains \Sexpr{nrow(rulexicon::hash_rusentilex_2016)} unique words or phrases, and
 \Sexpr{nrow(rulexicon::hash_rusentilex_2016[rulexicon::hash_rusentilex_2016$sentiment != "neutral",])} has non-neutral sentiment scores.
}
\details{
The lexicon contains the following types of sentiment-related words and phrases:
 \itemize{
   \item general sentiment (opinionated) words from domain-oriented Russian sentiment lists vocabularies;
   \item non-opinionated words with connotations usually convey information about negative or positive phenomena (facts) in
   social life;
   \item slang and curse words from Twitter.
 }
 The words in the lexicon having different sentiment orientations in specific senses are linked to appropriate concepts of
 the thesaurus of Russian language RuThes (\url{https://www.labinform.ru/pub/ruthes/}). All lexicon entries are classified
 according to four sentiment categories and three sources of sentiment (opinion, emotion, or fact).

Based on the RuSentiLex the polarity table with sentiment scores was generated by maintainer of \pkg{rulexicon} package,
see \code{\link{hash_sentiment_rusentilex_2016}}.
}
\section{License}{

According to information from Natalya Loukachevitch the lexicon RuSentiLex is published under
Creative Commons "Attribution-NonCommercial-ShareAlike" 4.0 International License (CC BY-NC-SA 4.0).
}

\references{
Loukachevitch N., Levchik A., 2016. Creating a General Russian Sentiment Lexicon. In Proceedings of Language Resources and Evaluation Conference LREC-2016. URL: \url{http://www.lrec-conf.org/proceedings/lrec2016/pdf/285_Paper.pdf}

RuSentiLex project web-page: \url{https://www.labinform.ru/pub/rusentilex/index.htm}
}
\keyword{datasets}
