% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hash_sentiment_ecsentithemelex.R
\docType{data}
\name{hash_sentiment_ecsentithemelex}
\alias{hash_sentiment_ecsentithemelex}
\title{Polarity Table of Sentiment-Thematic Dictionary of Economic Terms - EcSentiThemeLex}
\format{
A data table with \Sexpr{nrow(rulexicon::hash_sentiment_ecsentithemelex)} rows
and \Sexpr{ncol(rulexicon::hash_sentiment_ecsentithemelex)} variables:
\describe{
    \item{token}{a token (word)}
    \item{score}{a sentiment score: from −2 (strongly negative) to +2 (strongly positive) with step 1}
}
}
\usage{
data(hash_sentiment_ecsentithemelex)
}
\description{
The polarity table of sentiment-thematic dictionary of economic, financial and legal terms - EcSentiThemeLex.
 Table contains \Sexpr{nrow(rulexicon::hash_sentiment_ecsentithemelex)} words forms, and
 \Sexpr{nrow(rulexicon::hash_sentiment_ecsentithemelex[rulexicon::hash_sentiment_ecsentithemelex$score !=0,])}
 has non-neutral sentiment scores.
}
\details{
The polarity table was generated from original lexicon table (see \code{\link{key_ecsentithemelex}}) based on
the following rules:
 \itemize{
   \item tokens consisting of more than one word were discarded (as required for bag-of-words sentiment analysis)
   \item all possible word forms were added as separete \code{token}s
   \item for the words containing the letter "ё" the spelling option with letter "е" were added as separete \code{token}
 }
}
\section{License}{

The dictionary is published under Creative Commons "Attribution-NonCommercial-ShareAlike" 4.0 International License (CC BY-NC-SA 4.0).
For additional permissions (including the commercial use) please contact to Elena Fedorova <ecolena@mail.ru>.
}

\references{
Fedorova, E., Afanasyev, D., Demin, I., Lazarev, A., Nersesyan, R., Pyltsin, I.V. (2020).
Development of a tonal-thematic dictionary EcSentiThemeLex for the analysis of economic texts in Russian.
Journal of Applied Informatics, 6 (15), 58–77. DOI: \url{https://doi.org/10.37791/2687-0649-2020-15-6-58-77}.
}
\keyword{datasets}
