% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hash_valence_shifters_ru.R
\docType{data}
\name{hash_valence_shifters_ru}
\alias{hash_valence_shifters_ru}
\title{Russian Valence Shifters Table}
\format{
A data table with \Sexpr{nrow(rulexicon::hash_valence_shifters_ru)} rows
and \Sexpr{ncol(rulexicon::hash_valence_shifters_ru)} variables:
\describe{
    \item{x}{the textual token (word or phrase)}
    \item{y}{shifting coefficient}
    \item{t}{the type of valence shifter: 1 = negators, 2 = amplifiers, 3 = deamplifiers}
}
}
\source{
\code{\link{list_valence_shifters}}
}
\usage{
data(hash_valence_shifters_ru)
}
\description{
A table containing a valence shifters words that can alter a polarized word's meaning.
}
\details{
The dataset is based on \code{\link[sentometrics]{list_valence_shifters}} from \code{\link[sentometrics]{sentometrics}}
package (only English words) that was filtered and translated via Google Translate to Russian language.
}
\keyword{datasets}
