% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key_climate_change.R
\docType{data}
\name{key_climate_change}
\alias{key_climate_change}
\title{Dictionary of Climate Change Disclosure of Fedorova and Martynova}
\format{
A data frame with \Sexpr{nrow(rulexicon::key_climate_change)} rows
and \Sexpr{ncol(rulexicon::key_climate_change)} variables:
\describe{
    \item{token}{token (word or phrase)}
    \item{regex}{regular expression for all word forms of the \code{token} (including "е-ё" dualism)}
}
}
\source{
Intergovernmental Panel on Climate Change's Glossary: \url{https://www.ipcc.ch/site/assets/uploads/2018/03/wg2TARannexB.pdf}

World Bank Group's Glossary: \url{https://climateknowledgeportal.worldbank.org/media/document/CCKP_glossary.pdf}

Кокорин А.О., Липка О.Н., Суляндзига Р.В. Изменение климата. Глоссарий терминов, используемых в работе РКИК ООН. WWF России,
Москва, 2015 г., 92 с. URL: \url{https://wwf.ru/upload/iblock/509/wwf_glossary_web.pdf}.
}
\usage{
data(key_climate_change)
}
\description{
A dataset containing the dictionary (taxonomy) for climate change disclosures estimation based
 on three climate glossaries compiled and augmented by Fedorova E. and Martynova M.
 Table contains \Sexpr{nrow(rulexicon::key_climate_change)} words and phrases.
}
\details{
The dictionary is based on climate glossaries from three sources (Intergovernmental Panel on Climate Change,
World Bank Group and World Wildlife Fund) and augmented by Fedorova E. and Martynova M. It is additionally
supplemented by manual processing of climate change words from 235 sustainability and annual reports
of Russian companies. This dictionary enables to estimate the level of climate change disclosure
that may be useful for revealing the company’s concerns and actions over one of the thorniest issues in the society.
}
\section{License}{

The dictionary is published under Creative Commons "Attribution-NonCommercial-ShareAlike" 4.0 International License
(CC BY-NC-SA 4.0). For additional permissions (including the commercial use) please contact to Elena Fedorova <ecolena@mail.ru>.
}

\references{
Fedorova, E. and Martynova, M. (2021). Signalling and Legitimacy Theories for Explaining Climate Information Disclosure
by Russian Companies. Journal of Corporate Finance Research 15(2), p. 16-26.
DOI: \url{https://doi.org/10.17323/j.jcfr.2073-0438.15.2.2021.16-26}.
}
\keyword{datasets}
