% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key_legal_forms.R
\docType{data}
\name{key_legal_forms}
\alias{key_legal_forms}
\title{Firms Legal Forms Dictionary}
\format{A data frame with \Sexpr{nrow(rulexicon::key_legal_forms)} rows
and \Sexpr{ncol(rulexicon::key_legal_forms)} variables:
\describe{
    \item{abbreviation}{abbreviation of legal form name}
    \item{name}{legal form name}
    \item{regex}{regular expression for legal form abbreviation}
}}
\source{
\url{http://www.migrakvota.gov.ru/index.php?mod=class_show&act=5}
}
\usage{
data(key_legal_forms)
}
\description{
A dataset containing a dictionary of firms legal forms names and abbreviations.
}
\section{License}{

The dictionary is published under Public domain license.
}

\keyword{datasets}
