% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profanity_bally.R
\docType{data}
\name{profanity_bally}
\alias{profanity_bally}
\title{Profane Words List of bally_36 LiveJournal user}
\format{
A character vector with \Sexpr{length(rulexicon::profanity_bally)} elements.
}
\source{
Complex source, see References
}
\usage{
data(profanity_bally)
}
\description{
A dataset containing a character vector of profane words from LiveJournal user bally_36
 augmented by maintainer of \code{\link[rulexicon]{rulexicon}} package using additional list from www.spy-soft.net user IVO.
}
\section{License}{

There are no license informations provided by LiveJournal user bally_36 or www.spy-soft.net user IVO explicitly.
But content was published for free re-usage.
}

\references{
LiveJournal bally_36 users profane words list: \url{https://bally-36.livejournal.com/66906.html}

www.spy-soft.net IVO users profane words list: \url{http://www.spy-soft.net/slovar-dlya-bruta-russkij-mat}
}
\keyword{datasets}
