% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rulexicon-package.R
\docType{package}
\name{rulexicon}
\alias{rulexicon}
\alias{package-rulexicon}
\title{Russian Lexicons for Text Analysis}
\description{
An R data package containing a collection of Russian lexicons, polarity tables, words lists, lemmatization tables and etc. for text analysis.
}
