% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sw_nltk_ru.R
\docType{data}
\name{sw_nltk_ru}
\alias{sw_nltk_ru}
\title{Stopwrods List of Python's NLTK Library (augmented)}
\format{A character vector with \Sexpr{length(rulexicon::sw_nltk_ru)} elements.}
\source{
Complex source, see References
}
\usage{
data(sw_nltk_ru)
}
\description{
A dataset containing a character vector of stopwords based on Python's \code{NLTK} library augmented by Michail Kalinskiy using
 filtered \code{stopwords-ru} list.
}
\section{License}{

The Python's NLTK library and JavaScript's stopwords-iso NPM package stopwords lists are published under MIT License.
}

\references{
Python's NLTK library \code{stopwords} list: \url{https://github.com/mitmedialab/DataBasic/blob/master/nltk_data/corpora/stopwords/russian}

JavaScript's \code{stopwords-iso} NPM package stopwords list: \url{https://github.com/stopwords-iso/stopwords-ru}

Filtered by Michail Kalinskiy vesrion of JavaScript's \code{stopwords-iso} NPM package stopwords list:
\url{https://dev.kmint21.info/posts/python-summa}
}
\keyword{datasets}
