% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq_last_names.R
\docType{data}
\name{freq_last_names}
\alias{freq_last_names}
\title{Frequencies Table of Russian Last Names}
\format{A data table with \Sexpr{nrow(rulexicon::freq_last_names)} rows
and \Sexpr{ncol(rulexicon::freq_last_names)} variables:
\describe{
    \item{surname}{a last name}
    \item{freq}{the proportion relative to the most frequent surname}
}}
\source{
\url{http://hdl.handle.net/10995/1929}
}
\usage{
data(freq_last_names)
}
\description{
A dataset containing the most frequent Russian last names of Zhuravlev (2005).
}
\details{
The frequency table is based on the research results of Zhuravlev (2015). The table contains male surnames.
Most of them can be transformed into the female surnames by adding the ending "a" (but there are some exceptions).
In the future, it is planned to extend the table by including the female surnames in an explicit form.
}
\section{License}{
 There are no license informations provided by Zhuravlev A.F. explicitly. But the Problems of Onomastics journal is published
on an open access basis under the terms of the Creative Commons "Attribution-Non-Commercial-NoDerivs" 4.0 International License (CC BY-NC-ND 4.0).
}

\references{
Zhuravlev, A.F. (2005). On the Russian Family Names Statistics. Problems of Onomastics 2, pp. 126-146. URL: \url{http://hdl.handle.net/10995/1929}.
}
\keyword{datasets}
