% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key_corporate_report_standard.R
\docType{data}
\name{key_corporate_report_standard}
\alias{key_corporate_report_standard}
\title{Global Reporting Initiative Standards Dictionary augmented by Fedorova, Khrustova and Demin}
\format{A data frame with \Sexpr{nrow(rulexicon::key_corporate_report_standard)} rows
and \Sexpr{ncol(rulexicon::key_corporate_report_standard)} variables:
\describe{
    \item{order}{order number of section in report}
    \item{section}{name of report section}
    \item{main_token}{main or grouping token (term)}
    \item{token}{token (term), this can be equal to main_token}
    \item{regex}{regular expression for all word forms of the token}
}}
\source{
\url{https://www.globalreporting.org/standards/gri-standards-download-center/}
}
\usage{
data(key_corporate_report_standard)
}
\description{
A dataset containing Fedorova, Khrustova and Demin's corporate report standard dictionary: an economic taxonomy
 to verify compliance of public company reports with Global Reporting Initiative (GRI) Standards.
}
\details{
The dictionary is based on Global Reporting Initiative Standards glossary (\url{https://www.globalreporting.org/standards/})
augmented by Fedorova, Khrustova and Demin in the following:
\itemize{
   \item terms of general nature were deleted
   \item terms were extended and aligned with the terminology of Russian companies reports
   \item glossary were supplemented by additional topics and related terms: sponsorship, social responsibility, innovativeness, technology
   \item synonyms and different word forms were added as universal regular expression
 }
The dictionary can be used to verify compliance of public company reports with GRI Standards,
as well as for thematic analysis of company reports.
}
\section{License}{

The dictionary is published under Creative Commons "Attribution-NonCommercial-ShareAlike" 4.0 International License (CC BY-NC-SA 4.0).
For additional permissions (including for commercial applications) please contact to Elena Fedorova <ecolena@mail.ru>.
}

\references{
Fedorova, E.A. Khrustova, L.E. Demin I.S. (2020). The impact of the completeness of non-financial information disclosure by Russian companies on their investment attractiveness. Russian Management Journal 18(1).
}
\keyword{datasets}
